+++ server/src/domain/auth/dto/oauth-config.dto.ts
