+++ server/src/domain/auth/response-dto/oauth-config-response.dto.ts
