+++ server/src/domain/index.ts
@@ -13,7 +13,6 @@ export * from './facial-recognition';
 export * from './job';
 export * from './media';
 export * from './metadata';
-export * from './oauth';
 export * from './partner';
 export * from './person';
 export * from './search';
@@ -25,4 +24,3 @@ export * from './storage-template';
 export * from './system-config';
 export * from './tag';
 export * from './user';
-export * from './user-token';
