--- server/src/domain/oauth/dto/index.ts
@@ -1,2 +1,0 @@
-export * from './oauth-auth-code.dto';
-export * from './oauth-config.dto';
