--- server/src/domain/oauth/index.ts
@@ -1,4 +1,0 @@
-export * from './dto';
-export * from './oauth.constants';
-export * from './oauth.service';
-export * from './response-dto';
