--- server/src/domain/oauth/oauth.constants.ts
@@ -1,1 +1,0 @@
-export const MOBILE_REDIRECT = 'app.immich:/';
