--- server/src/domain/oauth/response-dto/index.ts
@@ -1,1 +1,0 @@
-export * from './oauth-config-response.dto';
