--- server/src/domain/user-token/index.ts
@@ -1,2 +1,0 @@
-export * from './user-token.core';
-export * from './user-token.repository';
