+++ server/src/infra/repositories/user-token.repository.ts
@@ -1,4 +1,4 @@
-import { IUserTokenRepository } from '@app/domain/user-token';
+import { IUserTokenRepository } from '@app/domain';
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
